Update JSON type and add sketch of control flow
The new JSON type should be ready for unmarshaling after some bug fixes,
especially concerning the weather list of descriptions.

The control flow is sketched out but requires implementation. One idea is to
issue an HTTP GET to the OpenWeatherData API, get a Response back, then read the
body of the response as []bytes. This can then be unmarshaled.