Smarter comparison to yesterday's temperature
This uses a better algorithm for comparing the current temperature to that of
yesterday: it takes the current Unix timestamp and subtracts 24 hours worth of
seconds. We then do a historical data query for that exact time. This should
give us at most two hours (either the previous or next hour) of readings.

A bonus is this allows for more precise descriptions (e.g. "this afternoon" as
opposed to only "today").